real-estate-agent/requirements.txt
fastapi
uvicorn
pydantic
pandas
faiss-cpu
sentence-transformers
rank_bm25
gunicorn
scikit-learn
transformers